# docs/openapi/service.openapi.yaml
openapi: 3.0.3
info:
  title: 小红书内容检索服务
  version: 0.1.0
  description: >
    基于 Spider_XHS 二次开发的小红书内容检索服务 OpenAPI 草案。
servers:
  - url: https://your-server.example.com
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
paths:
  /healthz:
    get:
      summary: 健康检查
      operationId: healthz
      responses:
        "200":
          description: 服务正常
  /api/v1/rednote/search:
    post:
      summary: 获取小红书搜索结果
      operationId: searchRedNote
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - keyword
                - cookie
              properties:
                keyword:
                  type: string
                note_type:
                  type: string
                  enum:
                    - default
                    - image
                    - video
                  default: default
                publish_time:
                  type: string
                  enum:
                    - default
                    - 1d
                    - 7d
                    - 180d
                  default: default
                sort_by:
                  type: string
                  enum:
                    - general
                    - latest
                    - most_liked
                    - most_commented
                    - most_collected
                  default: general
                page_count:
                  type: integer
                  minimum: 1
                  maximum: 10
                  default: 1
                cookie:
                  type: string
      responses:
        "200":
          description: 搜索响应
        "400":
          description: 非法请求
        "500":
          description: 内部错误或上游失败
  /api/v1/rednote/detail:
    post:
      summary: 通过 URL 获取小红书笔记详情
      operationId: getRedNoteDetail
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - url
                - cookie
              properties:
                url:
                  type: string
                  format: uri
                cookie:
                  type: string
      responses:
        "200":
          description: 详情响应
        "400":
          description: 非法请求
        "500":
          description: 内部错误或上游失败
  /api/v1/storage/notes:
    get:
      summary: 获取已落库的笔记列表
      operationId: listStoredNotes
      security:
        - ApiKeyAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: 落库笔记列表
  /api/v1/storage/notes/{note_id}:
    get:
      summary: 获取单条已落库笔记
      operationId: getStoredNote
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: note_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 单条笔记详情
        "404":
          description: 未找到记录
  /api/v1/storage/sync/pending:
    get:
      summary: 获取待同步任务
      operationId: listPendingSyncTasks
      security:
        - ApiKeyAuth: []
      parameters:
        - in: query
          name: target
          schema:
            type: string
            default: openclaw_bitable
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: 待同步任务列表
  /api/v1/storage/sync/tasks/{task_id}/success:
    post:
      summary: 标记同步任务成功
      operationId: markSyncTaskSuccess
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: task_id
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                bitable_record_id:
                  type: string
      responses:
        "200":
          description: 同步任务已标记成功
        "404":
          description: 未找到任务
  /api/v1/storage/sync/tasks/{task_id}/failed:
    post:
      summary: 标记同步任务失败
      operationId: markSyncTaskFailed
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: task_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - error_message
              properties:
                error_message:
                  type: string
      responses:
        "200":
          description: 同步任务已标记失败
        "404":
          description: 未找到任务